Add missing tests for region selector
Also add a minor fix to check on the workspace type in update_workspace
Rename the test file to be less generic

Re #34170
